Refactor Chat component to use custom hook for conversation reset
